+++ roxctl/central/debug/dump.go
@@ -31,7 +31,7 @@ func DumpCommand() *cobra.Command {
 		}),
 	}
 	flags.AddTimeoutWithDefault(c, dumpTimeout)
-	c.Flags().BoolVar(&withLogs, "logs", false, "logs=true will retrieve the dump without logs from Central")
+	c.Flags().BoolVar(&withLogs, "logs", false, "Include logs in Central dump")
 	c.PersistentFlags().StringVar(&outputDir, "output-dir", "", "output directory for bundle contents (default: auto-generated directory name inside the current directory)")
 
 	return c